ld symbol reference roadblock - relates to datatypes

+++ Makefile
@@ -2,6 +2,8 @@
 # Top-level build for rivermeta
 #
 
+include src/oscheck.mk
+
 BASE_DIR=.
 SRC_ROOT=./src
 
@@ -27,10 +29,15 @@ clean:
 	$(MAKE) --no-print-directory -C src clean
 	$(RM) waterslide waterslide-parallel $(QUIETOUT)
 	$(RM) wsalias wsman $(QUIETOUT)
-#	$(RM) $(WS_BIN_DIR)/* $(QUIETOUT)
-	$(RM) $(WS_PROCS_DIR)/proc_* $(QUIETOUT)
+ifndef ISWINDOWS
 	$(RM) $(WS_LIB_DIR)/wsdt_* $(QUIETOUT)
 	$(RM) $(WS_LIB_DIR)/libwaterslid*.a $(QUIETOUT)
+	$(RM) $(WS_PROCS_DIR)/proc_*
+else
+	@if EXIST lib $(RM) lib\wsdt_*
+	@if EXIST lib $(RM) lib\libwaterslide*.a
+	@if EXIST procs $(RM) procs\proc_*
+endif
 
 scour: clean
 	$(RM) $(WS_LIB_DIR) $(WS_PROCS_DIR) $(QUIETOUT)
